backend/requirements.txt
flask
flask-cors
sqlalchemy
pdfplumber
python-pptx
requests
bcrypt
pyjwt
